Moved reading the different versions to their own class
